+++ packages/encoding/package.json
@@ -39,5 +39,8 @@
   },
   "engines": {
     "node": ">=18.0.0"
+  },
+  "dependencies": {
+    "@noble/curves": "^1.3.0"
   }
 }
